refactoring: extract stdin and boltdb readers
